Update and/or add license notices to files
All files now have license notices!
