+++ src/main/pluto/commands/music/lyrics/handle.ts
@@ -1,7 +1,24 @@
+/*  This file is part of Pluto.
+ *  Copyright (C) 2021 Secnyt
+ *
+ *  Pluto is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 import MessageHandler from "../../../handlers/MessageHandler";
 import PlutoError from "../../../api/error/PlutoError";
 import * as solenolyrics from 'solenolyrics'
-import LyricsEmbed from "../../../embeds/LyricsEmbed";
+import LyricsEmbed from "./lib/LyricsEmbed";
 
 export default async function LyricsHandle (msg: any): Promise<PlutoError> {
     const content = MessageHandler.formattedAfterCommand(msg.content)
